requirements.txt
pymysql
pyaml